pyproject.toml
[tool.poetry]
name = "iris_backend"
version = "0.1.0"
description = ""
authors = ["Sasank <gssasank@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
app = "^0.0.1"
motor = "^3.0.0"
pytest = "^7.1.2"
pymongo = "^4.2.0"
mccabe = "^0.7.0"
dnspython = "^2.2.1"
email-validator = "^1.2.1"
python-dotenv = "^0.20.0"
fastapi-users = "^10.1.5"
spacy = "^3.4.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
